backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Firebase Admin SDK
firebase-admin==6.4.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-bigquery==3.17.0
google-cloud-storage==2.14.0

# JWT & Auth
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & CORS
httpx==0.26.0
aiohttp==3.9.1

# MailSlurp (NUEVO)
mailslurp-client==15.19.18

# Utils
python-dotenv==1.0.0
pandas==2.2.0

# Tableau (si usas Python SDK)
tableauserverclient==0.29